[REFACTOR] restructure canvas helper files
