+++ resources/js/service/canvas/blocks/LogoBlock.js
@@ -1,4 +1,4 @@
-import DrawBase from "../DrawBase";
+import DrawBase from "../misc/DrawBase";
 
 class LogoBlock extends DrawBase {
     constructor() {
