+++ resources/js/service/canvas/elements/background/Background.js
@@ -1,4 +1,4 @@
-import DrawBase from "../../DrawBase";
+import DrawBase from "../../misc/DrawBase";
 
 export default class Background extends DrawBase {
     set width(width) {
