+++ resources/js/service/canvas/misc/BorderHelper.js
